Add component for mobile market detail
